Copied updated data model from case-processor and return events associated with case

+++ src/main/java/uk/gov/ons/census/caseapisvc/service/CaseService.java
@@ -40,7 +40,7 @@ public class CaseService {
         .orElseThrow(() -> new CaseIdNotFoundException(caseIdUUID.toString()));
   }
 
-  public Case findByReference(long reference) {
+  public Case findByReference(int reference) {
     log.debug("Entering findByReference");
 
     return caseRepo
